Put billing unlinking code behind feature flag.

+++ api/src/main/java/org/pmiops/workbench/initialcredits/InitialCreditsService.java
@@ -499,12 +499,14 @@ public class InitialCreditsService {
     } catch (WorkbenchException e) {
       logger.error("Failed to delete apps and runtimes for workspace {}", namespace, e);
     }
-    try {
-      fireCloudService.removeBillingAccountFromBillingProjectAsService(namespace);
-      logger.info("Removed initial credits billing account from workspace {}", namespace);
-    } catch (WorkbenchException e) {
-      logger.error(
-          "Failed to remove initial credits billing account from workspace {}", namespace, e);
+    if (workbenchConfigProvider.get().featureFlags.enableUnlinkBillingForInitialCredits) {
+      try {
+        fireCloudService.removeBillingAccountFromBillingProjectAsService(namespace);
+        logger.info("Removed initial credits billing account from workspace {}", namespace);
+      } catch (WorkbenchException e) {
+        logger.error(
+            "Failed to remove initial credits billing account from workspace {}", namespace, e);
+      }
     }
   }
